refactor: extract useMarkdownMode + useContentDisplay composables (#507 step 5)
- useMarkdownMode: raw/rendered toggle + localStorage persistence
- useContentDisplay: derive isMarkdown/isHtml/isJson/isJsonl flags,
  sandboxedHtml, jsonTokens, jsonlLines, mdFrontmatter
- Drop now-unused imports (wrapHtmlWithPreviewCsp, tokenizeJson,
  tokenizeJsonl, prettyJson, extractFrontmatter) from FilesView
